buck2_futures

Module cancellation

Defines a future with explicit cancellation

Structs

CancellationContext

Context available to the function running inside the future to control and manage it’s own cancellation

CancellationHandle

A handle providing the ability to explicitly cancel the associated ExplicitlyCancellableFuture.

CancellationObserver

A Future that resolves only when the associated task has been canceled.

DisableCancellationGuard

A marker that indicates that cancellations have been disabled indefinitely for this task.

DropcancelHandle

A handle that will cancel the associated ExplicitlyCancellableFuture when it is dropped.

Enums

CriticalSectionGuard

When held, prevents cancellation of the current explicitly cancellable future. When no longer held (unless drop or cancellations are disabled was not ran), cancellations will be allowed. disable_cancellationsing this will leak the guard, causing cancellations to be disabled forever, regardless of whether or not the task is already cancelled.