buck2_futures::drop

Struct DropTogether

pub struct DropTogether<F, D> { /* private fields */ }

A future with a value attached that will be dropped together when the main future is dropped.

Implementations

impl<F, D> DropTogether<F, D>

pub fn new(fut: F, t: D) -> Self