buck2_futures::spawn

Function spawn_dropcancel

pub fn spawn_dropcancel<F, T, S>(
    f: F,
    spawner: &dyn Spawner<S>,
    ctx: &S,
) -> DropcancelJoinHandle<T> 
where
    for<'a> F: FnOnce(&'a CancellationContext) -> BoxFuture<'a, T> + Send,
    T: Any + Send + 'static,

Spawn a future and return a DropcancelJoinHandle. The future will begin execution even before the handle is polled. The future will be able to observe and control its cancellation with the provided CancellationContext.

When the handle is dropped, the task will be cancelled. The task can be detached by calling DropcancelJoinHandle::detach.