buck2_futures::spawner

Trait Spawner

pub trait Spawner<T>: Send + Sync {
    // Required method
    fn spawn(
        &self,
        ctx: &T,
        fut: BoxFuture<'static, Box<dyn Any + Send + 'static>>,
    ) -> JoinHandle<Box<dyn Any + Send + 'static>>;
}

A spawner that utilizes a generic context to decorate/wrap a future and spawn it. If a future needs to return a value, use a oneshot channel instead.

Required Methods

fn spawn( &self, ctx: &T, fut: BoxFuture<'static, Box<dyn Any + Send + 'static>>, ) -> JoinHandle<Box<dyn Any + Send + 'static>>